✨ refactor: JPQL 내 LIMIT 대신 Pageable 사용으로 수정 (#44)

+++ src/main/java/com/umc/momenty/domain/chat/service/query/ConversationQueryService.java
@@ -6,7 +6,7 @@ import java.util.List;
 import com.umc.momenty.domain.chat.dto.res.ChatResDTO;
 
 public interface ConversationQueryService {
-	List<ChatResDTO.ConversationListDTO> getConversationList(Long userId, Long cursorId, LocalDateTime cursorLastChatDate, Long count);
+	List<ChatResDTO.ConversationListDTO> getConversationList(Long userId, Long cursorId, LocalDateTime cursorLastChatDate, int count);
 
-	List<ChatResDTO.ConversationListDTO> searchConversationList(Long userId, String keyword, Long cursorId, LocalDateTime cursorLastChatDate, Long count);
+	List<ChatResDTO.ConversationListDTO> searchConversationList(Long userId, String keyword, Long cursorId, LocalDateTime cursorLastChatDate, int count);
 }
